[add] インターフェイス追加
