+++ Assets/Scripts/Runtime/1.Domain/OutGame/Scenario/AnimationEvent.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ac9b5a7ed563182468bf310622a60db8
