+++ Assets/Scripts/Runtime/1.Domain/OutGame/Scenario/BackgroundEvent.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 1ebdade2a4b66ea48a99766b2c0b4336
