+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IAnimationRepository.cs
@@ -1,0 +1,9 @@
+using KillChord.Runtime.Domain;
+
+namespace KillChord.Runtime.Application
+{
+    public interface IAnimationRepository
+    {
+        bool TryFindById(string id, out AnimationDefinition definition);
+    }
+}
