+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IAnimationRepository.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6374da37bd6a5a847bb06ebcce68eef0
