+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IBackgroundRepository.cs
@@ -6,9 +6,4 @@ namespace KillChord.Runtime.Application
     {
         bool TryFindById(string id, out BackgroundDefinition definition);
     }
-
-    public interface IAnimationRepository
-    {
-        bool TryFindById(string id, out AnimationDefinition definition);
-    }
 }
