+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IBackgroundRepository.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 02fcb247ef049fb4c8b5b7c62c875933
