--- Assets/Scripts/Runtime/2.Application/OutGame/Scenario/ICatalogRepositories.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 8de5d7ae575340249b9900a776078ecb
