+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IScenarioEventEmitter.cs
@@ -1,0 +1,11 @@
+using System.Threading;
+using System.Threading.Tasks;
+using KillChord.Runtime.Domain;
+
+namespace KillChord.Runtime.Application
+{
+    public interface IScenarioEventEmitter
+    {
+        ValueTask EmitAsync(IScenarioEvent scenarioEvent, CancellationToken ct);
+    }
+}
