+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IScenarioEventEmitter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e45f363b83216584fb18959df9fd8d93
