+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IScenarioEventHandler.Generic.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8717b6f213774c24da6e4be8a5f1d2ec
