+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IScenarioRepository.cs
@@ -1,0 +1,9 @@
+using KillChord.Runtime.Domain;
+
+namespace KillChord.Runtime.Application
+{
+    public interface IScenarioRepository
+    {
+        ScenarioData FindById(string id);
+    }
+}
