+++ Assets/Scripts/Runtime/2.Application/OutGame/Scenario/IScenarioRepository.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 308a593c171373d40b2adc6cee3b0247
